+++ tests/profanity.spec.ts
@@ -1,48 +1,9 @@
 import { expect } from "chai";
 
-import { profanity, Profanity, CensorType, ProfanityOptions } from "../src";
+import { Profanity, CensorType, ProfanityOptions } from "../src";
 
 describe("Profanity", () => {
   describe("Class instantiation", () => {
-    describe("ProfanityOptions", () => {
-      it("should create ProfanityOptions with default values", () => {
-        const options = new ProfanityOptions();
-        expect(options.wholeWord).to.be.true;
-        expect(options.grawlix).to.equal("@#$%&!");
-        expect(options.grawlixChar).to.equal("*");
-      });
-
-      it("should create ProfanityOptions with custom values", () => {
-        const options = new ProfanityOptions();
-        options.wholeWord = false;
-        options.grawlix = "***";
-        options.grawlixChar = "#";
-        expect(options.wholeWord).to.be.false;
-        expect(options.grawlix).to.equal("***");
-        expect(options.grawlixChar).to.equal("#");
-      });
-
-      it("should create ProfanityOptions with all partial custom values", () => {
-        const options = new ProfanityOptions({
-          wholeWord: false,
-          grawlix: "***",
-          grawlixChar: "#",
-        });
-        expect(options.wholeWord).to.be.false;
-        expect(options.grawlix).to.equal("***");
-        expect(options.grawlixChar).to.equal("#");
-      });
-
-      it("should create ProfanityOptions with some partial custom values", () => {
-        const options = new ProfanityOptions({
-          wholeWord: false,
-        });
-        expect(options.wholeWord).to.be.false;
-        expect(options.grawlix).to.equal("@#$%&!");
-        expect(options.grawlixChar).to.equal("*");
-      });
-    });
-
     describe("Profanity", () => {
       it("should create Profanity instance with default options", () => {
         const profanityInstance = new Profanity();
@@ -81,326 +42,6 @@ describe("Profanity", () => {
         expect(profanityInstance.options.wholeWord).to.be.false;
         expect(profanityInstance.options.grawlix).to.equal("@#$%&!");
         expect(profanityInstance.options.grawlixChar).to.equal("*");
-      });
-    });
-  });
-
-  describe("exists", () => {
-    describe("wholeWord = true", () => {
-      it("should return true when profanity exists in a sentence", () => {
-        expect(profanity.exists("I like big butts and I cannot lie")).to.be.true;
-      });
-
-      it("should return false when profanity is not a whole word in a sentence", () => {
-        expect(profanity.exists("Should we censor the word arsenic?")).to.be.false;
-      });
-
-      it("should return true when profanity exists within multiple lines", () => {
-        expect(
-          profanity.exists(`
-          Nothing profane on line 1.
-          Censoring butt on line 2.
-          Nothing profane on line 3.
-        `),
-        ).to.be.true;
-      });
-
-      it("should return false when profanity does not exist", () => {
-        expect(profanity.exists("I like big glutes and I cannot lie")).to.be.false;
-      });
-
-      it("should return true when profanity is surrounded by punctuation", () => {
-        expect(profanity.exists("What the (butt)!")).to.be.true;
-      });
-
-      it("should return false when profanity is part of a larger word", () => {
-        expect(profanity.exists("I'm feeling passionate today")).to.be.false;
-      });
-
-      it("should return true when profanity is at the beginning of a sentence", () => {
-        expect(profanity.exists("Butt is a profane word")).to.be.true;
-      });
-
-      it("should return true when profanity is at the end of a sentence", () => {
-        expect(profanity.exists("Don't be a butt.")).to.be.true;
-      });
-
-      it("should return false for words that are substrings of profane words", () => {
-        expect(profanity.exists("I need to assess the situation")).to.be.false;
-      });
-
-      it("should return true when profanity is separated by hyphens", () => {
-        expect(profanity.exists("Don't be a butt-head")).to.be.true;
-        expect(profanity.exists("Don't be a head-butt")).to.be.true;
-      });
-
-      it("should return true when profanity is separated by underscores", () => {
-        expect(profanity.exists("Don't be a butt_face")).to.be.true;
-        expect(profanity.exists("Don't be a face_butt")).to.be.true;
-      });
-
-      it("should return false when profanity is part of a word separated by a hyphen", () => {
-        expect(profanity.exists("Don't be an arsenic-head")).to.be.false;
-        expect(profanity.exists("Don't be a head-arsenic")).to.be.false;
-      });
-
-      it("should return false when profanity is part of a word separated by an underscore", () => {
-        expect(profanity.exists("Don't be an arsenic_head")).to.be.false;
-        expect(profanity.exists("Don't be a head_arsenic")).to.be.false;
-      });
-
-      it("should return false when profanity is part of a URL", () => {
-        expect(profanity.exists("Visit https://www.example.com/assets/image.jpg")).to.be.false;
-      });
-
-      it("should return true when profanity is surrounded by emoji", () => {
-        expect(profanity.exists("That's 💩butt💩")).to.be.true;
-      });
-    });
-
-    describe("wholeWord = false", () => {
-      let customProfanity: Profanity;
-
-      before(() => {
-        customProfanity = new Profanity({ wholeWord: false });
-      });
-
-      it("should return true when profanity is part of a word in a sentence", () => {
-        expect(customProfanity.exists("Should we censor the word arsenic?")).to.be.true;
-      });
-
-      it("should return false when profanity does not exist", () => {
-        expect(customProfanity.exists("I like big glutes and I cannot lie")).to.be.false;
-      });
-
-      it("Should return false when the last character is an 'A' with no profanity (A$$ edge case)", () => {
-        expect(customProfanity.exists("FUNTIMESA")).to.be.false;
-      });
-
-      it("Should return true when the last character is an 'A' and there is profanity (A$$ edge case)", () => {
-        expect(customProfanity.exists("BUTTSA")).to.be.true;
-      });
-
-      it("Should return true when some regex characters are present as profanity", () => {
-        expect(customProfanity.exists("lovea$$")).to.be.true;
-      });
-
-      it("should return true when profanity is at the beginning of a word", () => {
-        expect(customProfanity.exists("buttress the wall")).to.be.true;
-      });
-
-      it("should return true when profanity is at the end of a word", () => {
-        expect(customProfanity.exists("kickbutt performance")).to.be.true;
-      });
-
-      it("should return true when profanity is in the middle of a word", () => {
-        expect(customProfanity.exists("Massachusetts")).to.be.true;
-      });
-
-      it("should return true for words that are substrings of profane words", () => {
-        expect(customProfanity.exists("I need to assess the situation")).to.be.true;
-      });
-
-      it("should return true when profanity is part of a URL", () => {
-        expect(customProfanity.exists("Visit https://www.example.com/assets/image.jpg")).to.be.true;
-      });
-
-      it("should return true when profanity is separated by hyphens", () => {
-        expect(customProfanity.exists("Don't be a butt-head")).to.be.true;
-        expect(customProfanity.exists("Don't be a head-butt")).to.be.true;
-        expect(customProfanity.exists("Don't be an arsenic-head")).to.be.true;
-        expect(customProfanity.exists("Don't be a head-arsenic")).to.be.true;
-      });
-
-      it("should return true when profanity is separated by underscores", () => {
-        expect(customProfanity.exists("Don't be a butt_head")).to.be.true;
-        expect(customProfanity.exists("Don't be a head_butt")).to.be.true;
-        expect(customProfanity.exists("Don't be an arsenic_head")).to.be.true;
-        expect(customProfanity.exists("Don't be a head_arsenic")).to.be.true;
-      });
-
-      it("should return true when profanity is surrounded by emoji", () => {
-        expect(customProfanity.exists("That's 💩butt💩")).to.be.true;
-      });
-    });
-
-    describe("Case sensitivity", () => {
-      it("should detect mixed case profanity", () => {
-        expect(profanity.exists("Don't be a BuTt")).to.be.true;
-      });
-
-      it("should detect all uppercase profanity", () => {
-        expect(profanity.exists("DON'T BE A BUTT")).to.be.true;
-      });
-
-      it("should detect all lowercase profanity", () => {
-        expect(profanity.exists("don't be a butt")).to.be.true;
-      });
-
-      it("should detect profanity with alternating case", () => {
-        expect(profanity.exists("dOn'T bE a BuTt")).to.be.true;
-      });
-
-      it("should detect profanity with random casing", () => {
-        expect(profanity.exists("DoN't Be A bUtT")).to.be.true;
-      });
-    });
-  });
-
-  describe("censor", () => {
-    describe("Default censoring (CensorType.Word)", () => {
-      it("should replace multiple profane words within a sentence with grawlix", () => {
-        const censored = profanity.censor("I like big butts (aka arses) and I cannot lie");
-        expect(censored).to.equal(`I like big ${profanity.options.grawlix} (aka ${profanity.options.grawlix}) and I cannot lie`);
-      });
-
-      it("should replace profane words within a multi-line sentence with grawlix", () => {
-        const censored = profanity.censor(`
-          Nothing profane on line 1.
-          Censoring butt on line 2.
-          Nothing profane on line 3.
-        `);
-        expect(censored).to.equal(`
-          Nothing profane on line 1.
-          Censoring ${profanity.options.grawlix} on line 2.
-          Nothing profane on line 3.
-        `);
-      });
-
-      it("sentences without profanity should not be altered", () => {
-        const original = "I like big glutes and I cannot lie";
-        expect(profanity.censor(original)).to.equal(original);
-      });
-
-      it("should censor profanity at the beginning of a sentence", () => {
-        expect(profanity.censor("Butt is a profane word")).to.equal(`${profanity.options.grawlix} is a profane word`);
-      });
-
-      it("should censor profanity at the end of a sentence", () => {
-        expect(profanity.censor("Don't be a butt")).to.equal(`Don't be a ${profanity.options.grawlix}`);
-      });
-
-      it("should censor multiple occurrences of the same profane word", () => {
-        expect(profanity.censor("Butt, butt, butt!")).to.equal(
-          `${profanity.options.grawlix}, ${profanity.options.grawlix}, ${profanity.options.grawlix}!`,
-        );
-      });
-
-      it("should not censor parts of words that contain profanity", () => {
-        expect(profanity.censor("I need to assess the situation")).to.equal("I need to assess the situation");
-      });
-
-      it("should censor profanity separated by hyphens", () => {
-        expect(profanity.censor("Don't be a butt-head")).to.equal(`Don't be a ${profanity.options.grawlix}-head`);
-      });
-
-      it("should censor profanity separated by underscores", () => {
-        expect(profanity.censor("Don't be a butt_head")).to.equal(`Don't be a ${profanity.options.grawlix}_head`);
-      });
-
-      it("should censor profanity surrounded by emoji", () => {
-        expect(profanity.censor("That's 💩butt💩")).to.equal(`That's 💩${profanity.options.grawlix}💩`);
-      });
-    });
-
-    describe("CensorType.FirstChar", () => {
-      it("should replace first character of each profane word with grawlix character", () => {
-        const censored = profanity.censor("I like big butts (aka arses) and I cannot lie", CensorType.FirstChar);
-        expect(censored).to.equal(`I like big ${profanity.options.grawlixChar}utts (aka ${profanity.options.grawlixChar}rses) and I cannot lie`);
-      });
-
-      it("should preserve case when censoring first character", () => {
-        expect(profanity.censor("Don't be a BuTt", CensorType.FirstChar)).to.equal(`Don't be a ${profanity.options.grawlixChar}uTt`);
-      });
-
-      it("should censor first character of profanity at the beginning of a sentence", () => {
-        expect(profanity.censor("Butt is a profane word", CensorType.FirstChar)).to.equal(`${profanity.options.grawlixChar}utt is a profane word`);
-      });
-
-      it("should censor first character of profanity at the end of a sentence", () => {
-        expect(profanity.censor("Don't be a butt.", CensorType.FirstChar)).to.equal(`Don't be a ${profanity.options.grawlixChar}utt.`);
-      });
-
-      it("should censor first character of profanity separated by hyphens", () => {
-        expect(profanity.censor("Don't be a butt-head", CensorType.FirstChar)).to.equal(`Don't be a ${profanity.options.grawlixChar}utt-head`);
-      });
-
-      it("should censor first character of profanity separated by underscores", () => {
-        expect(profanity.censor("Don't be a butt_head", CensorType.FirstChar)).to.equal(`Don't be a ${profanity.options.grawlixChar}utt_head`);
-      });
-    });
-
-    describe("CensorType.FirstVowel", () => {
-      it("should replace first vowel of each profane word with grawlix character", () => {
-        const censored = profanity.censor("I like big butts (aka arses) and I cannot lie", CensorType.FirstVowel);
-        expect(censored).to.equal(`I like big b${profanity.options.grawlixChar}tts (aka ${profanity.options.grawlixChar}rses) and I cannot lie`);
-      });
-
-      it("should not censor if no vowels are present", () => {
-        expect(profanity.censor("tsk tsk", CensorType.FirstVowel)).to.equal("tsk tsk");
-      });
-
-      it("should censor first vowel of profanity at the beginning of a sentence", () => {
-        expect(profanity.censor("Butt is a profane word", CensorType.FirstVowel)).to.equal(`B${profanity.options.grawlixChar}tt is a profane word`);
-      });
-
-      it("should censor first vowel of profanity at the end of a sentence", () => {
-        expect(profanity.censor("Don't be a butt.", CensorType.FirstVowel)).to.equal(`Don't be a b${profanity.options.grawlixChar}tt.`);
-      });
-
-      it("should handle profane words with no vowels", () => {
-        expect(profanity.censor("Don't say tsk", CensorType.FirstVowel)).to.equal("Don't say tsk");
-      });
-
-      it("should censor first vowel of profanity separated by hyphens", () => {
-        expect(profanity.censor("Don't be a butt-head", CensorType.FirstVowel)).to.equal(`Don't be a b${profanity.options.grawlixChar}tt-head`);
-      });
-
-      it("should censor first vowel of profanity separated by underscores", () => {
-        expect(profanity.censor("Don't be a butt_head", CensorType.FirstVowel)).to.equal(`Don't be a b${profanity.options.grawlixChar}tt_head`);
-      });
-    });
-
-    describe("CensorType.AllVowels", () => {
-      it("should replace all vowels within each profane word with grawlix character", () => {
-        const censored = profanity.censor("I like big butts (aka arses) and I cannot lie", CensorType.AllVowels);
-        expect(censored).to.equal(
-          `I like big b${profanity.options.grawlixChar}tts (aka ${profanity.options.grawlixChar}rs${profanity.options.grawlixChar}s) and I cannot lie`,
-        );
-      });
-
-      it("should preserve case when censoring all vowels", () => {
-        expect(profanity.censor("BuTt", CensorType.AllVowels)).to.equal(`B${profanity.options.grawlixChar}Tt`);
-      });
-
-      it("should censor all vowels of profanity at the beginning of a sentence", () => {
-        expect(profanity.censor("Butt is a profane word", CensorType.AllVowels)).to.equal(`B${profanity.options.grawlixChar}tt is a profane word`);
-      });
-
-      it("should censor all vowels of profanity at the end of a sentence", () => {
-        expect(profanity.censor("Don't be a butt.", CensorType.AllVowels)).to.equal(`Don't be a b${profanity.options.grawlixChar}tt.`);
-      });
-
-      it("should handle profane words with no vowels", () => {
-        expect(profanity.censor("Don't say tsk", CensorType.AllVowels)).to.equal("Don't say tsk");
-      });
-    });
-
-    describe("Case sensitivity", () => {
-      it("should censor while preserving case", () => {
-        expect(profanity.censor("Don't be a BuTt")).to.equal("Don't be a @#$%&!");
-      });
-
-      it("should censor all uppercase profanity", () => {
-        expect(profanity.censor("DON'T BE A BUTT")).to.equal("DON'T BE A @#$%&!");
-      });
-
-      it("should censor mixed case profanity", () => {
-        expect(profanity.censor("Don't Be A bUtT")).to.equal("Don't Be A @#$%&!");
-      });
-
-      it("should censor profanity with alternating case", () => {
-        expect(profanity.censor("dOn'T bE a BuTt")).to.equal("dOn'T bE a @#$%&!");
       });
     });
   });
@@ -585,34 +226,6 @@ describe("Profanity", () => {
         const customProfanity = new Profanity({ grawlixChar: "X" });
         expect(customProfanity.censor("You're a butt", CensorType.FirstChar)).to.equal("You're a Xutt");
       });
-    });
-  });
-
-  describe("Multi-word profanities", () => {
-    it("should detect multi-word profanities", () => {
-      expect(profanity.exists("He's a fudge packer")).to.be.true;
-      expect(profanity.exists("That's a blow job")).to.be.true;
-      expect(profanity.exists("Don't be a son-of-a-bitch")).to.be.true;
-    });
-
-    it("should censor multi-word profanities", () => {
-      expect(profanity.censor("He's a fudge packer")).to.equal(`He's a ${profanity.options.grawlix}`);
-      expect(profanity.censor("That's a blow job")).to.equal(`That's a ${profanity.options.grawlix}`);
-      expect(profanity.censor("Don't be a son-of-a-bitch")).to.equal(`Don't be a ${profanity.options.grawlix}`);
-    });
-
-    it("should not detect partial matches of multi-word profanities", () => {
-      expect(profanity.exists("I like to pack fudge for desserts")).to.be.false;
-      expect(profanity.exists("The wind blew jobs away")).to.be.false;
-      expect(profanity.exists("He's the son of a businessman")).to.be.false;
-    });
-
-    it("should handle multi-word profanities with different censor types", () => {
-      expect(profanity.censor("He's a fudge packer", CensorType.FirstChar)).to.equal(`He's a ${profanity.options.grawlixChar}udge packer`);
-      expect(profanity.censor("That's a blow job", CensorType.FirstVowel)).to.equal(`That's a bl${profanity.options.grawlixChar}w job`);
-      expect(profanity.censor("Don't be a son-of-a-bitch", CensorType.AllVowels)).to.equal(
-        `Don't be a s${profanity.options.grawlixChar}n-${profanity.options.grawlixChar}f-${profanity.options.grawlixChar}-b${profanity.options.grawlixChar}tch`,
-      );
     });
   });
 });
